refactor(aws): avoid using iam_role data source
Our strategy to use a data source was to access the ARN of the
generated IAM Role, or the provided IAM Role name if the user
decides not to create it. This caused confusions since the data
resource will generate a constant change while running terraform
apply, so we are trying not to use this pattern anymore.

Example of a second terraform apply:
```
  ~ resource "aws_iam_policy" "cross_account_policy" {
      ~ policy      = jsonencode(...) -> (known after apply)
  }

  ~ resource "lacework_integration_aws_cfg" "default" {
      ~ credentials {
            external_id = "..."
          ~ role_arn    = "..." -> (known after apply)
      }
  }

Plan: 0 to add, 2 to change, 0 to destroy.

Do you want to perform these actions?
  Terraform will perform the actions described above.
  Only 'yes' will be accepted to approve.

  Enter a value: yes

  ...

Apply complete! Resources: 0 added, 0 changed, 0 destroyed.
```

BREAKING-CHANGE This might be a breaking change for our current users!

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>