fix(gcp): for project level integrations
This change is fixing multiple things for project level integrations.

Additionally, we are now using the data source
`google_storage_project_service_account` that will access the storage
service account and therefore, Google will create the right service
account for us. If we don't ask for it, Google doens't create it by
default, more docs at:

=> https://cloud.google.com/storage/docs/projects#service-accounts

If the service account is not there, we could add a local-exec
provisioner to call an API that is documeted at:

=> https://cloud.google.com/storage-transfer/docs/reference/rest/v1/googleServiceAccounts/get

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>