fix(azure): attribute reference causing owner to not be identified (#122)
We were trying to run this Script as the administrator/owner of the Azure tenant/subscription and we encountered errors. We had to switch the reference to be `.id` instead of `.objectId` as the `objectId` key did not exist and the value `null` was getting assigned when performing the Owner permissions User check. When we changed it to `.id`, the script began to function as we expected.